+++ crates/containerdebug/CHANGELOG.md
@@ -7,10 +7,15 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Increased the default `--loop` interval from every minute to every 30 minutes ([#23]).
-- Collect and output the open files limit ([#45])
+- Collect and output the open files limit ([#45]).
+
+### Fixes
+
+- Move the span inside the loop ([#46]).
 
 [#23]: https://github.com/stackabletech/containerdebug/pull/23
 [#45]: https://github.com/stackabletech/containerdebug/pull/45
+[#46]: https://github.com/stackabletech/containerdebug/pull/46
 
 ## [0.1.1] - 2024-12-16